// src/api/models/CrearCitaAdministradorRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CrearCitaAdministradorRequest = {
    clienteId: number;
    fechaHoraInicio: string;
    tipoMantenimiento: string;
    observaciones?: string | null;
};